+++ Emulator/Agnus/DmaDebugger.cpp
@@ -12,7 +12,7 @@
 DmaDebugger::DmaDebugger(Amiga &ref) : AmigaComponent(ref)
 {
     // By default all DMA channels are visualized, except the CPU channel
-    for (unsigned i = 0; i < BUS_COUNT; i++) {
+    for (isize i = 0; i < BUS_COUNT; i++) {
         visualize[i] = (i != BUS_NONE) && (i != BUS_CPU);
     }
